storing LP address in depositor; cleanup

+++ stride-covenant/packages/depositor/src/msg.rs
@@ -12,7 +12,7 @@ pub struct InstantiateMsg {
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 pub struct WeightedReceiver {
-    pub amount: u128,
+    pub amount: i64,
     pub address: String,
 }
